Added tests, and docstrings, ready for the first round of review
